Tests/LibWeb/Text/expected/cross-origin-isolated-worker-inherits.txt

window crossOriginIsolated: true
worker crossOriginIsolated: true
worker received SharedArrayBuffer: true
